A mechanism for dislaying Parser objects' help messages.
Parsers now have an attribute called `help`, which should be a callable
that prints the help messages.

`__init__` can now take a class like HelpFormatter as an argument; it
creates an instance of that class, with itself as an argument, and then
saves it as `self.help`.

Also modified `from_function`, so we can pass along kwargs.