Make new arguments work in the parser.
Use arg.underscored consistently in when parsing. Before any flags would
give errors because it would be checking that the given flags are a
subset of the parser's possible flags, and the parser's possible flags
were a list of Parser objects rather than a list of strings.

It would also feed the wrong things to the functions.